requirements.txt
Flask==1.1.1
pymongo==3.10.0
azure-storage-file==2.1.0
rq==1.1.0
Pillow==6.2.1
Flask-Caching==1.8.0
